# src/main/java/com/augusto/login_system/user/dto/ChangePasswordRequest.java
package com.augusto.login_system.user.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record ChangePasswordRequest(
        @NotBlank String currentPassword,
        @NotBlank @Size(min = 8) String newPassword,
        @NotBlank String confirmNewPassword
) {}